requirements.txt
asgiref==3.6.0
cloudinary==1.32.0
dj-database-url==0.5.0
Django==3.2.18
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
